requirements.txt
django
django-cors-headers
djangorestframework
psycopg2
PyJWT